Assignment 2 Design a Three-Layered ANN Classifier
You are provided with a bank customers dataset (Churn_Modelling.csv) with about 10,000 customer information that can be used to decide whether the customer is likely to churn. There are multiple features. Identify which features are significant in determining whether the customer will churn. The last column, “Exited,” tells whether the customer stayed with the bank (Exited = 0) or left the bank (Exited = 1). Write a python code to design a three-layered ANN classifier that can predict whether the customer will churn for the test data set, which is 20% of the total dataset. Print the confusion matrix and accuracy, and then, submit the python code.

Be sure to encode the categorical data and perform the feature scaling. Use ‘relu’ activation for the first and second layers and ‘sigmoid’ for the last dense layer. For compiling, use ‘adam’ optimizer; and loss should be ‘binary_crossentropy’ as this is a binary classification problem.